+++ CHANGELOG.md
@@ -40,6 +40,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   - `client.batch.updateByFilter(filter, criteriaList, options)`
   - `client.batch.hisDeleteByFilter(filter, start, end, options)`
   - `client.batch.migrateHistory(fromEntity, toEntity)`
+  - `client.batch.addChildrenByFilter(filter, children, tagMap)`
 - Added new utility class `EntityCriteria`  to be used with `client.batch.updateByFilter`.
 - Added new function `entityCount(filter)` to get the number of entities from a filter via a GraphQL query.
 - Added new function `findAsId(filter, limit)` to optimise functions that only require the ids of the entity,